refactor: use daemon sdk noop reconciler
